+++ HISTORY.rst
@@ -5,6 +5,9 @@ Version: 2.0.0 - (IN DEVELOPMENT)
 
 **Enhancements**
 
+* #625: exposed a new ``FTPHandler.encoding`` attribute defaulting to
+  ``'utf-8'``. It can be used to change the encoding used for client / server
+  communication.
 * #629: removed Python 2.7 support.
 * #637: remove copies of asyncore.py and asynchat.py. Use backports from PYPI
   instead.  (patch by @penguinpee)